feat(script): rename project

+++ script/rename-project.sh
@@ -2,9 +2,16 @@
 
 # Fail on any error
 set -euo pipefail
+#set -o xtrace
 
+PROJECT_ROOT_NAME=sea
 OLD_PROJECT_NAME=$1
 NEW_PROJECT_NAME=$2
+OLD_PROJECT_DASH_NAME="${PROJECT_ROOT_NAME}-${OLD_PROJECT_NAME}"
+NEW_PROJECT_DASH_NAME="${PROJECT_ROOT_NAME}-${NEW_PROJECT_NAME}"
+OLD_PROJECT_JOINT_NAME="${PROJECT_ROOT_NAME}${OLD_PROJECT_NAME}"
+NEW_PROJECT_JOINT_NAME="${PROJECT_ROOT_NAME}${NEW_PROJECT_NAME}"
+
 OLD_GIT_REPOSITORY_NAME=$3 #github.com/kaydxh
 NEW_GIT_REPOSITORY_NAME=$4 #git.code.oa.com/kaydxh
 
@@ -27,29 +34,42 @@ function checkParams {
 # %/*代表取从头到最后一个slash之前的所有内容
 # */代表去取从第一个slash之后的所有内容
 
+# sed 默认大小写敏感，/i参数不区分大小写
 function renameFilesAndDirectories {
   for it in $(find . ! -path "*third_party*" ! -path "*node_modules*" -type f -path "*${OLD_PROJECT_NAME}*")
   do
-    oldFile=${it}
-    newFile=`echo ${it} | sed -e "s/${OLD_PROJECT_NAME}/${NEW_PROJECT_NAME}/g"`
-
-   newDir=${newFile%/*}
-   mkdir -p ${newDir}; mv  ${oldFile} ${newFile}
+    renameProjectDir ${it} ${OLD_PROJECT_DASH_NAME} ${NEW_PROJECT_DASH_NAME}
+    renameProjectDir ${it} ${OLD_PROJECT_JOINT_NAME} ${NEW_PROJECT_JOINT_NAME}
+    renameProjectDir ${it} ${OLD_PROJECT_NAME} ${NEW_PROJECT_NAME}
   done
 
   rmDirectories 
+}
+
+# $1 file or dir name
+# 老服务文件名/目录名 替换为 新服务文件名/目录名
+function renameProjectDir {
+    oldFile=$1
+    #newFile=`echo ${oldFile} | sed -e "s/${OLD_PROJECT_NAME}/${NEW_PROJECT_NAME}/g"`
+    newFile=`echo ${oldFile} | sed -e "s/$2/$3/g"`
+    newDir=${newFile%/*}
+    if [[ -f ${oldFile} ]]; then mkdir -p ${newDir}; mv  ${oldFile} ${newFile}; fi
 }
 
 function rmDirectories {
   for it in $(find . ! -path "*third_party*" ! -path "*node_modules*"  -type d -path "*${OLD_PROJECT_NAME}*")
   do
     oldDir=${it}
+    echo "${oldDir}"
     rm -rf ${oldDir}
   done
 }
 
 function replaceContentOfFiles {
-  for it in $(grep -RIl --exclude-dir={*third_party*,*node_modules*,*output*} ${OLD_PROJECT_NAME} .)
+  UPPER_BEGIN_OLD_PROJECT_NAME=$(echo ${OLD_PROJECT_NAME:0:1} | tr '[a-z]' '[A-Z]')${OLD_PROJECT_NAME:1}
+  UPPER_BEGIN_NEW_PROJECT_NAME=$(echo ${NEW_PROJECT_NAME:0:1} | tr '[a-z]' '[A-Z]')${NEW_PROJECT_NAME:1}
+  #for it in $(grep -E -RIl --exclude-dir={*third_party*,*node_modules*,*output*} '${OLD_PROJECT_NAME}|${UPPER_BEGIN_OLD_PROJECT_NAME}' .)
+  for it in $(grep -E -RIl --exclude-dir={*third_party*,*node_modules*,*output*} "${OLD_PROJECT_NAME}|${UPPER_BEGIN_OLD_PROJECT_NAME}" .)
   do
     echo "${it}"
     # skip soft link file
@@ -57,17 +77,19 @@ function replaceContentOfFiles {
       continue
     fi
 
-   sed -i "" "s/${OLD_PROJECT_NAME}/${NEW_PROJECT_NAME}/g" "${it}"
+   sed -i "" "/Validate/!s/${OLD_PROJECT_NAME}/${NEW_PROJECT_NAME}/g" "${it}" 
 
-   UPPER_BEGIN_OLD_PROJECT_NAME=$(echo ${OLD_PROJECT_NAME:0:1} | tr '[a-z]' '[A-Z]')${OLD_PROJECT_NAME:1}
-   UPPER_BEGIN_NEW_PROJECT_NAME=$(echo ${NEW_PROJECT_NAME:0:1} | tr '[a-z]' '[A-Z]')${NEW_PROJECT_NAME:1}
+   #首字母大写替换
+   #将OLD_PROJECT_NAME变量值的首字母转化为大写，并保存在UPPER_BEGIN_OLD_PROJECT_NAME变量中
+   #UPPER_BEGIN_OLD_PROJECT_NAME=$(echo ${OLD_PROJECT_NAME:0:1} | tr '[a-z]' '[A-Z]')${OLD_PROJECT_NAME:1}
+   #UPPER_BEGIN_NEW_PROJECT_NAME=$(echo ${NEW_PROJECT_NAME:0:1} | tr '[a-z]' '[A-Z]')${NEW_PROJECT_NAME:1}
    sed -i "" "s/${UPPER_BEGIN_OLD_PROJECT_NAME}/${UPPER_BEGIN_NEW_PROJECT_NAME}/g" "${it}"
 
    # replace git name
-   sed -i "" "s/${UPPER_BEGIN_OLD_PROJECT_NAME}/${UPPER_BEGIN_NEW_PROJECT_NAME}/g" "${it}"
+   sed -i "" "s/${UPPER_BEGIN_OLD_PROJECT_NAME} /${UPPER_BEGIN_NEW_PROJECT_NAME}/g" "${it}"
 
    # support by base 4.0
-   # sed -i "" "s/${OLD_PROJECT_NAME^}/${NEW_PROJECT_NAME^}/g" "${it}"
+   #sed -i "" "s/${OLD_PROJECT_NAME^}/${NEW_PROJECT_NAME^}/g" "${it}"
   done
 }
 
@@ -81,16 +103,21 @@ function replaceGitRespositoryNameOfFiles {
     fi
 
     OLD_GIT_NAME=${OLD_GIT_REPOSITORY_NAME}/${OLD_PROJECT_NAME}
-    ESCAPED_OLD_GIT_NAME=$(printf '%s\n' "$OLD_GIT_NAME" | sed -e 's/[]\/$*.^[]/\\&/g')
     NEW_GIT_NAME=${NEW_GIT_REPOSITORY_NAME}/${NEW_PROJECT_NAME}
-    ESCAPED_NEW_GIT_NAME=$(printf '%s\n' "$NEW_GIT_NAME" | sed -e 's/[]\/$*.^[]/\\&/g')
-
-   sed -i "" "s/${ESCAPED_OLD_GIT_NAME}/${ESCAPED_NEW_GIT_NAME}/g" "${it}"
-
-   # support by base 4.0
-   # sed -i "" "s/${OLD_PROJECT_NAME^}/${NEW_PROJECT_NAME^}/g" "${it}"
+    replaceString ${OLD_GIT_NAME} ${NEW_GIT_NAME}
+    replaceString ${OLD_GIT_REPOSITORY_NAME} ${NEW_GIT_REPOSITORY_NAME}
   done
 }
+
+
+function replaceString {
+    ESCAPED_OLD_GIT_NAME=$(printf '%s\n' "$1" | sed -e 's/[]\/$*.^[]/\\&/g')
+    ESCAPED_NEW_GIT_NAME=$(printf '%s\n' "$2" | sed -e 's/[]\/$*.^[]/\\&/g')
+    sed -i "" "s/${ESCAPED_OLD_GIT_NAME}/${ESCAPED_NEW_GIT_NAME}/g" "${it}"
+   # support by base 4.0
+   #sed -i "" "s/${OLD_PROJECT_NAME^}/${NEW_PROJECT_NAME^}/g" "${it}"
+}
+
 
 function replaceProjectRootName {
   if [[ -d ../${OLD_PROJECT_NAME} ]]; then
@@ -98,9 +125,9 @@ function replaceProjectRootName {
   fi
 }
 
-#checkParams 
-#renameFilesAndDirectories 
-#replaceContentOfFiles 
-#replaceProjectRootName 
+checkParams 
+renameFilesAndDirectories
+replaceContentOfFiles 
+replaceProjectRootName 
 replaceGitRespositoryNameOfFiles
